This changes the fan component from using a hard coded dp for the fan speed.  The correct dp for fan speed is passed in when the component is referenced.  It also attempts to deal with fans that return 'slow', 'medium' and 'high' for fan speeds, instead of a number.
Fan controllers with more than 3 speeds are not supported in Home Assistant at this time; they
will simply have 3 speeds and no more.  There is a discussion of changing this to a percentage
and letting the component map to the proper setting.

Fan controllers with an 'auto' setting is also not really supported. I simply mapped it to
the 'medium' speed.

This includes an attempt to deal with oscillate setting stack dumping; it is still hardcoded to dps 8, where
now if dp 8 is not defined, it ignores it; on one of my fans, 8 is valid, but it's not the oscillate setting.

This also includes a fix for when a sensor value is text.

I am using this with  air purifier and a ceiling fan controller with dimmer.

The ceiling fan controller is the Treatlife Single Pole WiFi light switch/Fan speed controller.
The air purifier is a RENPHO Air Purifier.