Updates fan controller.
This adds support for 3 speed fans, 4 speeds fan (with choice of 3 of the 4 speeds),
fans speed of 1,2,3,4 or low,medium,mid,high, and auto.

Fans that have variable speeds, or greater than 4 speeds is not supported
at this time.  HA does not support variable speed fans and more than 3 speeds.