localtuya: refactor discovery processing
If the discovered device_id has no configuration we can stop
immediately.  On first discovery add it to the cache.

Do not replace the user-provided hostname with another identifier
(such as an IP address), as that's not what the user will be looking
for in log messages that expose CONF_HOST.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>