Show warning instead of Error exception trace in log
Shows only a warning in the log files if a device is offline.
This is a fairly common situation for light switches, and it
previously showed an unreadable exception as an Error.
Now it will just show a simple warning in the logs when
localtuya failed to reach the device.

Fixes #650