Region WE and EA
As tuya have introduced some new datacenters the url-scheme implemented broke. It is now instead using that as a fallback and a if statement for the we and ea datacenters is adddded to cloud_api. It also updates the config flow to allow for the new regions to be selected.